Rename com.github.walfie.granblue -> walfie.gbf
Should do this sooner than later. The name was so long.
