+++ build.sbt
@@ -8,7 +8,7 @@ lazy val commonSettings = Seq(
 lazy val core = (project in file("core"))
   .settings(commonSettings: _*)
   .settings(
-    name := "granblue-raid-finder-core",
+    name := "gbf-raidfinder-core",
     libraryDependencies ++= Seq(
       "com.typesafe.akka" %% "akka-agent" % Versions.Akka,
       "io.monix" %% "monix" % Versions.Monix,
@@ -19,7 +19,7 @@ lazy val core = (project in file("core"))
   )
 
 lazy val protocol = (crossProject.crossType(CrossType.Pure) in file("protocol"))
-  .settings(name := "granblue-raid-finder-protocol")
+  .settings(name := "gbf-raidfinder-protocol")
   .enablePlugins(ScalaJSPlugin)
   .settings(commonSettings: _*)
   .settings(ScalaPB.settings: _*)
@@ -29,7 +29,7 @@ lazy val protocolJS = protocol.js
 lazy val server = (project in file("server"))
   .settings(commonSettings: _*)
   .settings(
-    name := "granblue-raid-finder-server",
+    name := "gbf-raidfinder-server",
     libraryDependencies ++= Seq(
       "com.trueaccord.scalapb" %% "scalapb-json4s" % Versions.ScalaPB_json4s,
       "com.typesafe.play" %% "play-netty-server" % Versions.Play
@@ -41,7 +41,7 @@ lazy val client = (project in file("client"))
   .enablePlugins(ScalaJSPlugin)
   .settings(commonSettings: _*)
   .settings(
-    name := "granblue-raid-finder-client",
+    name := "gbf-raidfinder-client",
     persistLauncher in Compile := true,
     libraryDependencies ++= Seq(
       "org.scala-js" %%% "scalajs-dom" % "0.9.0"
