+++ client/src/main/resources/index.html
@@ -5,8 +5,9 @@
   <title></title>
 </head>
 <body>
-  <script src="../granblue-raid-finder-client-jsdeps.js"></script>
-  <script src="../granblue-raid-finder-client-fastopt.js"></script>
-  <script src="../granblue-raid-finder-client-launcher.js"></script>
+  <script src="../gbf-raidfinder-client-jsdeps.js"></script>
+  <script src="../gbf-raidfinder-client-fastopt.js"></script>
+  <script src="../gbf-raidfinder-client-launcher.js"></script>
 </body>
 </html>
+
