+++ client/src/main/scala/walfie/gbf/raidfinder/client/Application.scala
@@ -1,0 +1,45 @@
+package walfie.gbf.raidfinder.client
+
+import java.nio.ByteBuffer
+import org.scalajs.dom
+import scala.scalajs.js
+import walfie.gbf.raidfinder.protocol._
+import walfie.gbf.raidfinder.protocol.implicits._
+
+import js.JSConverters._
+import js.typedarray._
+import js.typedarray.TypedArrayBufferOps.bufferOps
+import js.{ArrayOps, JSApp}
+
+object Application extends JSApp {
+  def main(): Unit = {
+    val msg = RequestMessage().withRaidBossesMessage(RaidBossesRequest())
+
+    val socket = new dom.WebSocket("ws://localhost:9000/ws/raids", js.Array("binary"))
+    socket.binaryType = "arraybuffer"
+
+    val bytes = {
+      TypedArrayBuffer.wrap(new Int8Array(msg.toByteArray.toJSArray)).arrayBuffer
+    }
+
+    socket.onopen = { e: dom.Event =>
+      socket.send(bytes)
+    }
+
+    socket.onmessage = { e: dom.MessageEvent =>
+      val data = e.data match {
+        case buffer: ArrayBuffer =>
+          new Int8Array(buffer).toArray
+      }
+      val message = ResponseMessage.parseFrom(data) // TODO: validate
+
+      message.toResponse.foreach {
+        case r: RaidBossesResponse =>
+          r.raidBosses.foreach { rb =>
+            println(rb.bossName)
+          }
+      }
+    }
+  }
+}
+
