--- core/src/main/scala/com/github/walfie/granblue/raidfinder/package.scala
@@ -1,10 +1,0 @@
-package com.github.walfie.granblue
-
-import com.github.walfie.granblue.raidfinder.domain._
-import com.github.walfie.granblue.raidfinder.util.{ObservablesPartitioner, CachedObservablesPartitioner}
-
-package object raidfinder {
-  type RaidTweetsPartitioner = ObservablesPartitioner[BossName, RaidTweet]
-  type CachedRaidTweetsPartitioner = CachedObservablesPartitioner[BossName, RaidTweet]
-}
-
