+++ core/src/main/scala/walfie/gbf/raidfinder/StatusParser.scala
@@ -1,0 +1,41 @@
+package walfie.gbf.raidfinder
+
+import twitter4j._
+import walfie.gbf.raidfinder.domain._
+
+object StatusParser {
+  /** Regex to match Japanese raid request tweets */
+  val RaidRegex = "(.*)参加者募集！参戦ID：([0-9A-F]+)\n(.+)\n?.*".r
+
+  /** The source value for the official Granblue Twitter app */
+  val GranblueSource =
+    """<a href="http://granbluefantasy.jp/" rel="nofollow">グランブルー ファンタジー</a>"""
+
+  def parse(status: Status): Option[RaidInfo] = status.getText match {
+    case _ if status.getSource != GranblueSource => None
+
+    case RaidRegex(extraText, raidId, boss) =>
+      val bossName = boss.trim
+
+      val raidTweet = RaidTweet(
+        tweetId = status.getId,
+        screenName = status.getUser.getScreenName,
+        bossName = bossName,
+        raidId = raidId.trim,
+        profileImage = status.getUser.getProfileImageURLHttps,
+        text = extraText.trim,
+        createdAt = status.getCreatedAt
+      )
+
+      val image = getImageFromStatus(status)
+
+      Some(RaidInfo(raidTweet, image))
+
+    case _ => None
+  }
+
+  private def getImageFromStatus(status: Status): Option[RaidImage] = {
+    status.getMediaEntities.headOption.map(_.getMediaURLHttps)
+  }
+}
+
