+++ core/src/test/scala/walfie/gbf/raidfinder/util/TestObserver.scala
@@ -1,0 +1,25 @@
+package walfie.gbf.raidfinder.util
+
+import monix.execution.Ack
+import monix.reactive.Observer
+import scala.concurrent.Future
+
+class TestObserver[T] extends Observer[T] {
+  var received = Vector.empty[T]
+  var errorsReceived = Vector.empty[Throwable]
+  var isComplete: Boolean = false
+
+  def onNext(elem: T): Future[Ack] = {
+    received = received :+ elem
+    Ack.Continue
+  }
+
+  def onError(e: Throwable): Unit = {
+    errorsReceived = errorsReceived :+ e
+  }
+
+  def onComplete(): Unit = {
+    isComplete = true
+  }
+}
+
